+++ CHANGELOG.md
@@ -28,6 +28,9 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - added `Bar.getSystemZonedBeginTime`: the bar's begin time converted to system time zone
 - added `Bar.getSystemZonedEndTime`: the bar's end time converted to system time zone
 - added `BarSeries.getSeriesPeriodDescriptionInSystemTimeZone`: with times printed in system's default time zone
+- added `KRIIndicator`
+- Added constructor with `amount` for  `EnterAndHoldCriterion`
+- Added constructor with `amount` for  `VersusEnterAndHoldCriterion`
 
 ## 0.17 (released September 9, 2024)